Updated Search Algorithm (BUGGY)
If the user types an existing entry name on the search bar and presses the Enter key, they will be redirected to that entry's page. However, if the user doesn't type anything on the search bar, enters to the '/query' URL, or types a non-existent entry name, an error message will appear. I need to fix that bug.