Updated Search Algorithm
Now, if the user presses the Enter key while not typing anything on the search bar, or if they type a non-existent entry name on it, they will be redirected to a generic HTML error message that I created. Inside that 'if' statement, I will show the list of entries that are similar to whatever the user typed on the search bar.