BUG FIX: Search Algorithm (BUGGY)
I had to rewrite huge chunks of the code to make the search queries case insensitive. Now, the user's not forced to type anything in uppercase if the title of the entry is in uppercase to display it in the list of search results. However, I will have a bug which is preventing me to show a message that says 'sorry, but there are no results' if the user types a query that's not a substring of any existing entry title.