Added Create Page
I added a file called create.html, which has the page that will let the user create an entry for the wiki. If the user clicks on the link on the sidebar that says 'Create New Page', they will be taken to an empty page, which is the creation page.