Finished 4.a) from Create Algorithm
Although I still need to fix the styles of the form for the entry creation page, I have technically finished part 4.a) from the Create algorithm. Now, if the user enters to the creation page, a form with 2 inputs and a submit button will show up. And if you hit 'Submit', no error will show up.