Updated Creation Algorithm (BUGGY)
Users can now create entries and save them to the website. However, if the user types a '#' alongside the title, the user won't be able to reach that entry unless they use the 'Random page' link. Also, the title of the newly created entries is not being rendered once the user enters the page of the newly created entry.