BUG FIX: Updated Creation Algorithm
Turns out that the bug that gave me troubles with the '#', and the markdown <h1> titles not being rendered was caused because the '#' should be typed on the 'description' field of the form, NOT on the 'title' form. Also, I need to type the title twice (see CSS.md for a reference): once in the 'title' field, nad once in the 1st line of the 'description' field.