Updated Edition Algorithm
I added the edition page by creating a file called edit.html. I also edited the code in urls.py and on views.py so that, if the user types '/wiki/entry_title/edit' in the URL page, they will be taken into the edit page, where a form with a <textarea> tag will be pre-populated with the description of that entry.