Updated Edition Algorithm
I finished the 5.a) part of the algorithm. Now, if the user enters into an entry page, they will see a link that says 'Edit entry'. If they click on that link, they will be taken into an edition page. In this page, you will see a <textarea> field that's pre-populated with the description of that particular entry.